+++ flotils/__init__.py
@@ -9,10 +9,10 @@ An utility package
 
 __author__ = "the01"
 __email__ = "jungflor@gmail.com"
-__copyright__ = "Copyright (C) 2013-17, Florian JUNG"
+__copyright__ = "Copyright (C) 2013-18, Florian JUNG"
 __license__ = "MIT"
-__version__ = "0.3.5a0"
-__date__ = "2017-03-06"
+__version__ = "0.4.1"
+__date__ = "2018-01-27"
 
 import logging
 
@@ -20,7 +20,13 @@ from .loadable import Loadable, load_file, save_file, \
     loadJSON, saveJSON, loadJSONFile, saveJSONFile
 from .logable import Logable, ModuleLogable, get_logger
 from .runable import StartStopable, Startable, Stopable, StartException
+from .convenience import FromToDictBase, PrintableBase
 
 
 logger = logging.getLogger(__name__)
-__all__ = ["logable", "loadable", "runable"]
+__all__ = [
+    "logable", "Logable", "ModuleLogable", "get_logger",
+    "loadable", "Loadable", "load_file", "save_file",
+    "runable", "Startable", "StartException", "Stopable", "StartStopable",
+    "convenience", "FromToDictBase", "PrintableBase"
+]